// src/types/workout.ts
export interface Workout {
  id: number | string;
  name: string;
  image: string;
  muscleGroups: string[];
  equipment: string;
  difficulty: string;
  duration: number; 
  caloriesBurned: number; 
  sets: number;
  reps: string;
  rating: number;
  description: string;
  instructions: string[];
}